fix: umd namespacing
