+++ package.json
@@ -40,6 +40,7 @@
     "@znemz/js-common-prettierrc-clone": "^0.0.23",
     "@znemz/react-extras-jest": "^1.0.0-alpha.5",
     "bluebird": "^3.5.5",
+    "change-case": "^3.1.0",
     "coveralls": "^3.0.4",
     "del": "^4.1.1",
     "esm": "^3.2.22",
